Adding few edits.

+++ Week 50/464854/turn2_modelB_sample1.go
@@ -8,5 +8,5 @@ import (
 func main() {
 	var m runtime.MemStats
 	runtime.ReadMemStats(&m)
-	fmt.Println("Total Memory Allocated: ", m.HeapAlloc/1024/1024) // convert to MB()
+	fmt.Println("Total Memory Allocated: ", float64(float64(m.HeapAlloc/1024)/1024)) // convert to MB()
 }
